Develop

+++ enternal-rune-backend/src/main/java/iuh/fit/se/enternalrunebackend/util/GenerateQRURL.java
@@ -3,9 +3,11 @@ package iuh.fit.se.enternalrunebackend.util;
 import iuh.fit.se.enternalrunebackend.config.QRConfig;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class GenerateQRURL {
-    public String getQRURL(QRConfig qrConfig, double amount, String description) {
+    public String getQRURL(QRConfig qrConfig, BigDecimal amount, String description) {
         String acc =  qrConfig.getAcc();
         String bank = qrConfig.getBank();
         String template = qrConfig.getTemplate();
